requirements.txt
flask
gunicorn
scikit-learn
pandas
joblib
pytz
pytesseract
Pillow
requests
beautifulsoup4